pyproject.toml
[project]
name = "starter-english-teacher-agent"
version = "0.1.0"
description = "Agente professor de inglês com RAG usando o livro Starter da SOS English"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "ollama>=0.6.1",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
]

[tool.uv]
dev-dependencies = ["ruff"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
